Add puzzle generation and interactive curses game
The backtracking added a few commits back is a bit buggy - the generator
uses it extensively, and will fail frequently on a bad assert in
backtrack.
Puzzle generation needs some work, as the puzzles being generated are
simply not that hard. Further investigation is required to determine how
to fix this problem.
The curses game is fleshed out, but still needs a help file, and a way
of reversing a move. (One can currently pencil and ink in cells, use
auto-pencil, and view cells by number available. Autopencil should also
be optional.)